% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_function.R
\name{add_function}
\alias{add_function}
\title{add funtions that you want to store them in a function.R file}
\usage{
add_function(..., source_fun = F)
}
\arguments{
\item{virable}{param my_functions input any functions that you want to}
}
\value{
information about what did the function do
}
\description{
today,I create my fourth function,a very useful function.
}
\details{
you can use this function to receive any funtions that you want to store them in a function.R file
and source them automatically into your environment.
}
\examples{
add_function(source_fun = TRUE)
}
\keyword{add_function}
