% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages_version.R
\name{check_packages_version}
\alias{check_packages_version}
\title{check_packages_version}
\usage{
check_packages_version(packages = packages)
}
\arguments{
\item{vector}{contains package names used}
}
\value{
data.frame contain packages version info
}
\description{
output packages version info
}
\details{
you can use this function to receive a vector contained packages used
in your program.
}
\examples{
check_packages_version(c("maftools", "sad"))
}
