% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_installed_packages.R
\name{make_installed_packages}
\alias{make_installed_packages}
\title{Title make_installed_packages}
\usage{
make_installed_packages()
}
\value{
a file
}
\description{
Title make_installed_packages
}
\examples{
make_installed_packages()
}
