% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ppt_template.R
\name{make_ppt_template}
\alias{make_ppt_template}
\title{make_ppt_template}
\usage{
make_ppt_template(wd = work_folder, fn = file_name)
}
\arguments{
\item{fn}{}
}
\value{
folder
}
\description{
make_ppt_template
}
