% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rmd_template.R
\name{make_rmd_template}
\alias{make_rmd_template}
\title{Make own rmd file}
\usage{
make_rmd_template(
  filename = "file",
  project = "Work on the project",
  author = "Liang",
  dir = getwd()
)
}
\arguments{
\item{filename}{file name}

\item{project}{project name}

\item{author}{author}

\item{dir}{used_folder}
}
\value{
file name
}
\description{
Make own rmd file
}
\examples{
make_rmd_template("my_rmd", "project", "Liang")
}
