% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rprofile.R
\name{make_rprofile}
\alias{make_rprofile}
\title{Title make r profile}
\usage{
make_rprofile(new_file = ".Rprofile")
}
\arguments{
\item{new_file}{file name}
}
\value{
file name
}
\description{
Title make r profile
}
\examples{
make_rprofile(new_file=".Rprofile")
}
