% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ttoolkit}
\alias{ttoolkit}
\title{ttoolkit: A package for store my useful functions.}
\description{
The ttoolkit package provides three cool functions:
boost_install_packages, set_mirror and make_r_template.
}
\references{
Full documentation and demos: \url{https://github.com/mugpeng/ttoolkit};
}
